# Optimize CrlCollectionAccessor.Contains for large lists